services/insights-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
sumy==0.11.0
scikit-learn==1.5.0
numpy==1.26.4
nltk==3.8.1
python-dotenv==1.0.1
